support read and handle few of files

+++ beijing45/beijing54/CalculateModel/BasicDef/ScanFreqMonitorFile.h
@@ -223,8 +223,8 @@ struct ValueInTime
 	ngtime tTime;
 	float* pdLevelValue;
 	long nValueNumber;
-	float lon;
-	float lat;
+	double lon;
+	double lat;
 	//float height;
 	ValueInTime();
 	ValueInTime(long nValueNumber);
